requirements.txt
asgiref==3.2.3
boto3==1.12.5
botocore==1.15.5
dj-database-url==0.5.0
Django==3.0.2
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-heroku==0.3.1
django-markdown==0.8.4
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
